azure_score/requirements.txt
azure-cognitiveservices-speech>=1.43.0 